Adding service mocks and service unit tests

+++ internal/handler/author.go
@@ -3,7 +3,6 @@ package handler
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/murat96k/kitaptar.kz/api"
-	"log"
 	"net/http"
 )
 
@@ -12,8 +11,8 @@ func (h *Handler) getAllAuthors(ctx *gin.Context) {
 	authors, err := h.srvs.GetAllAuthors(ctx)
 
 	if err != nil {
-		log.Printf("Handler all authors getting error %w", err)
-		ctx.JSON(http.StatusInternalServerError, err)
+		ctx.AbortWithStatusJSON(http.StatusInternalServerError, api.Error{Message: err.Error()})
+		return
 	}
 
 	ctx.JSON(http.StatusOK, authors)
@@ -22,11 +21,15 @@ func (h *Handler) getAllAuthors(ctx *gin.Context) {
 func (h *Handler) getAuthorById(ctx *gin.Context) {
 
 	authorId := ctx.Param("id")
-	author, err := h.srvs.GetAuthorById(ctx, authorId)
+	if authorId == "" {
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, api.Error{Message: "author id is empty"})
+		return
+	}
 
+	author, err := h.srvs.GetAuthorById(ctx, authorId)
 	if err != nil {
-		log.Printf("Handler author getting by id error %w", err)
-		ctx.JSON(http.StatusNotFound, err)
+		ctx.AbortWithStatusJSON(http.StatusInternalServerError, api.Error{Message: err.Error()})
+		return
 	}
 
 	ctx.JSON(http.StatusOK, author)
@@ -36,14 +39,14 @@ func (h *Handler) createAuthor(ctx *gin.Context) {
 	var req api.AuthorRequest
 
 	if err := ctx.BindJSON(&req); err != nil {
-		log.Println("Bind json error ", err.Error())
-		ctx.JSON(http.StatusBadRequest, err)
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, api.Error{Message: err.Error()})
+		return
 	}
 
 	err := h.srvs.CreateAuthor(ctx, &req)
 	if err != nil {
-		log.Printf("Error %w", err)
-		ctx.JSON(http.StatusInternalServerError, err)
+		ctx.AbortWithStatusJSON(http.StatusInternalServerError, api.Error{Message: err.Error()})
+		return
 	}
 
 	ctx.JSON(http.StatusCreated, req)
@@ -53,19 +56,20 @@ func (h *Handler) updateAuthor(ctx *gin.Context) {
 	var req api.AuthorRequest
 
 	if err := ctx.BindJSON(&req); err != nil {
-		log.Println("Bind json error ", err.Error())
-		ctx.JSON(http.StatusBadRequest, err)
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, api.Error{Message: err.Error()})
+		return
 	}
+
 	authorId := ctx.Param("id")
-	//_, err := h.srvs.GetBookById(ctx, authorId)
-	//
-	//if err != nil {
-	//	log.Printf("Handler author getting by id error %w", err)
-	//}
+	if authorId == "" {
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, api.Error{Message: "author id is empty"})
+		return
+	}
+
 	err := h.srvs.UpdateAuthor(ctx, authorId, &req)
 	if err != nil {
-		log.Printf("Error %w", err)
-		ctx.JSON(http.StatusInternalServerError, err)
+		ctx.AbortWithStatusJSON(http.StatusInternalServerError, api.Error{Message: err.Error()})
+		return
 	}
 
 	ctx.JSON(http.StatusOK, req)
@@ -74,11 +78,16 @@ func (h *Handler) updateAuthor(ctx *gin.Context) {
 func (h *Handler) deleteAuthor(ctx *gin.Context) {
 
 	authorId := ctx.Param("id")
+	if authorId == "" {
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, api.Error{Message: "author id is empty"})
+		return
+	}
 
 	err := h.srvs.DeleteAuthor(ctx, authorId)
 	if err != nil {
-		ctx.JSON(http.StatusInternalServerError, err)
+		ctx.AbortWithStatusJSON(http.StatusInternalServerError, api.Error{Message: err.Error()})
+		return
 	}
 
-	ctx.JSON(http.StatusOK, "Author deleted")
+	ctx.JSON(http.StatusOK, api.Response{Message: "Author deleted"})
 }
